+++ src/vapi/insight/raw_client.py
@@ -158,9 +158,6 @@ class RawInsightClient:
             json=convert_and_respect_annotation_metadata(
                 object_=request, annotation=InsightControllerCreateRequest, direction="write"
             ),
-            headers={
-                "content-type": "application/json",
-            },
             request_options=request_options,
             omit=OMIT,
         )
@@ -406,9 +403,6 @@ class RawInsightClient:
             json=convert_and_respect_annotation_metadata(
                 object_=request, annotation=InsightControllerPreviewRequest, direction="write"
             ),
-            headers={
-                "content-type": "application/json",
-            },
             request_options=request_options,
             omit=OMIT,
         )
@@ -559,9 +553,6 @@ class AsyncRawInsightClient:
             json=convert_and_respect_annotation_metadata(
                 object_=request, annotation=InsightControllerCreateRequest, direction="write"
             ),
-            headers={
-                "content-type": "application/json",
-            },
             request_options=request_options,
             omit=OMIT,
         )
@@ -807,9 +798,6 @@ class AsyncRawInsightClient:
             json=convert_and_respect_annotation_metadata(
                 object_=request, annotation=InsightControllerPreviewRequest, direction="write"
             ),
-            headers={
-                "content-type": "application/json",
-            },
             request_options=request_options,
             omit=OMIT,
         )